Auth_app/requirements.txt
fastapi
uvicorn
sqlalchemy
bcrypt
pyjwt
pytest